Explicitly query controller mappings env var instead of hardcoding
Normally, SDL is supposed to look at the SDL_GAMECONTROLLERCONFIG_FILE
env var during SDL_Init(). But this doesn't seem to happen on all systems,
so to make sure we have mappings, explicitly query the env var ourselfes.

This replaces the hardcoded path, making it easier to change the path
without having to recompile the text viewer.